requirements.txt
langchain
langchain-community
dashscope
python-dotenv
pydantic
